test: collapse search helper suites
